streamlit/requirements.txt
streamlit
numpy==1.23.5
tensorflow==2.7.0
protobuf==3.20
Pillow==8.4.0
opencv-python-headless
joblib==1.1.0
psycopg2-binary==2.9.3
pip==23.2.1
scikit-learn
requests-toolbelt